feat(dsg): create new component LogInForm/ update index.html path
